+++ JsonToolsNppPlugin/JSONTools/RemesPath.cs
@@ -936,11 +936,12 @@ namespace JSON_Tools.JSON_Tools
             {
                 JNode arg = func.args[ii + 1];
                 if (arg is CurJson) { other_callables = true; }
-                argsCanBeFunctions[ii] = ((func.function.TypeOptions(ii + 1) & Dtype.FUNCTION) != 0)
-                    && !(func.function.conditionalExecution && ii == func.args.Count - 2); // if conditional execution, last arg is CurJson
-                                                                                           // that must be called before argfunction is evaluated
+                argsCanBeFunctions[ii] = (func.function.TypeOptions(ii + 1) & Dtype.FUNCTION) != 0;
                 other_args.Add(arg);
             }
+            if (func.function.conditionalExecution)
+                argsCanBeFunctions[argsCanBeFunctions.Length - 1] = false; // if conditional execution, last arg is CurJson
+                                                                           // that must be called before argfunction is evaluated
             Dtype out_type = func.function.OutputType(x);
             List<JNode> all_args = new List<JNode>(func.args.Count);
             foreach (var a in func.args)